LayoutTests/imported/w3c/web-platform-tests/IndexedDB/key-conversion-exceptions-expected.txt


FAIL IDBFactory cmp() static with throwing/invalid keys assert_throws: first key conversion with invalid key should throw DataError function "() => {
            receiver[method](invalid_key, throwing_key('getter 2'));
        }" threw object "getter 2: throwing from getter" that is not a DOMException DataError: property "code" is equal to undefined, expected 0
FAIL IDBCursor continue() method with throwing/invalid keys assert_throws: key conversion with invalid key should throw DataError function "() => {
            receiver[method](invalid_key);
        }" threw object "InvalidStateError (DOM IDBDatabase Exception 11): Failed to execute 'continue' on 'IDBCursor': The cursor is being iterated or has iterated past its end." that is not a DOMException DataError: property "code" is equal to 11, expected 0
PASS IndexedDB: Exceptions thrown during key conversion 
PASS IDBCursor update() method with throwing/invalid keys 
PASS IDBKeyRange only() static with throwing/invalid keys 
PASS IDBKeyRange lowerBound() static with throwing/invalid keys 
PASS IDBKeyRange upperBound() static with throwing/invalid keys 
FAIL IDBKeyRange bound() static with throwing/invalid keys assert_throws: first key conversion with invalid key should throw DataError function "() => {
            receiver[method](invalid_key, throwing_key('getter 2'));
        }" threw object "getter 2: throwing from getter" that is not a DOMException DataError: property "code" is equal to undefined, expected 0
FAIL IDBObjectStore add() method with throwing/invalid keys assert_throws: key conversion with throwing getter should rethrow function "() => {
            out_of_line[method]('value', throwing_key('getter'));
        }" threw object "DataCloneError (DOM IDBDatabase Exception 25): Failed to store record in an IDBObjectStore: An object could not be cloned." ("DataCloneError") expected object "[object Object]" ("getter")
FAIL IDBObjectStore put() method with throwing/invalid keys assert_throws: key conversion with throwing getter should rethrow function "() => {
            out_of_line[method]('value', throwing_key('getter'));
        }" threw object "DataCloneError (DOM IDBDatabase Exception 25): Failed to store record in an IDBObjectStore: An object could not be cloned." ("DataCloneError") expected object "[object Object]" ("getter")
PASS IDBObjectStore delete() method with throwing/invalid keys 
PASS IDBObjectStore get() method with throwing/invalid keys 
PASS IDBObjectStore getKey() method with throwing/invalid keys 
PASS IDBObjectStore getAll() method with throwing/invalid keys 
PASS IDBObjectStore getAllKeys() method with throwing/invalid keys 
PASS IDBObjectStore count() method with throwing/invalid keys 
PASS IDBObjectStore openCursor() method with throwing/invalid keys 
PASS IDBObjectStore openKeyCursor() method with throwing/invalid keys 
PASS IDBIndex get() method with throwing/invalid keys 
PASS IDBIndex getKey() method with throwing/invalid keys 
PASS IDBIndex getAll() method with throwing/invalid keys 
PASS IDBIndex getAllKeys() method with throwing/invalid keys 
PASS IDBIndex count() method with throwing/invalid keys 
PASS IDBIndex openCursor() method with throwing/invalid keys 
PASS IDBIndex openKeyCursor() method with throwing/invalid keys 
